requirements.txt
gTTS==2.2.3
Pillow==9.2.0
pytesseract==0.3.9
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
streamlit==1.10.0
translate==3.6.1
